refactor #34 : 기존 사이드바로직 total-evaluation폴더로 병합
