+++ src/features/total-evaluation/components/left-slide-panel-toggle/left-slide-panel-toggle.styles.ts
