--- src/features/total-evaluation/index.styles.ts
@@ -1,8 +1,0 @@
-import { css } from '@emotion/react';
-
-export const totalEvaluationPage = css`
-  display: flex;
-  flex-direction: column;
-  gap: 50px;
-  width: 100%;
-`;
