+++ src/features/total-evaluation/total-evaluation-page.styles.ts
@@ -6,8 +6,15 @@ export const container = css`
   display: flex;
 `;
 
-export const contents = (isOpen: boolean) => css`
+export const totalEvaluation = (isOpen: boolean) => css`
   flex: 1;
   margin-left: ${isOpen ? '300px' : '0'};
   transition: margin-left 0.2s ease-in-out;
+`;
+
+export const totalEvaluationSection = css`
+  display: flex;
+  flex-direction: column;
+  gap: 50px;
+  width: 100%;
 `;
