+++ src/features/total-evaluation/types/sidebar-Info-types.ts
