+++ src/features/total-evaluation/utils/adapt-accordion-format.ts
